Better parser - Still don't like it
The parser I created is not my favorite. It does not allow for sentences with extra words. It is built to understand sentences like "throw goblin wall," but not "throw goblin at the wall."

My goal is to create a parser that looks for the verb at the beginning of the sentence, the subject in the middle, and the item at the end. The parser should be capable of understanding sentences that do not include items.