Parser - Latest Edition
Current parser looks for a verb, and two components. The way it searches for a component in the room was annoying to figure out. By using a lambda function to look at each individual component in the array of components, we can compare the component names with the words in the input.